kbuild: deb-pkg: Add additional build dependencies
Currently there are a couple of build dependencies missing and, whilst they
will almost always be present on a developers workstation, these missing
dependencies will cause containerized builds to fail.

linux-libc-dev is needed since it no longer pulled in by default by the
debian build tools (and is needed to compile some of the host tools).

rsync is needed since it is used by the header install rules.

Signed-off-by: Daniel Thompson <daniel.thompson@linaro.org>