NOUPSTREAM: kbuild: deb-pkg: Meta-package support
Kernel packages are "special" because they are designed to be parallel
installed and therefore have a revision number in the package name.
That means we need to create a meta-package that depends on the latest
numbered version of the kernel. Without this apt will not automatically
keep up to date with the latest kernel.

Signed-off-by: Daniel Thompson <daniel.thompson@linaro.org>