NOUPSTREAM: Force C-A72 workarounds to deploy on C-A53
Currently we work around a bootloader bug on Pinebook Pro by disabling
the A72 cores at boot. This has a nasty side effect, which is that the
A72 errata logic doesn't deploy and the errata system refuses to allow
the A72 cores to come online.

Workaround this with simply brute force (hence the NOUPSTREAM).

Signed-off-by: Daniel Thompson <daniel.thompson@linaro.org>